+++ CHANGELOG.md
@@ -77,6 +77,7 @@
 - Introduce mayManageMailboxes() permissions. Removed the entry from the menu if permission is false. !1244 @jofranz
 - Moved database access from RegionXhr, MailboxXhr, and ForumService to gateways !1228 #9 @alex.simm
 - Removed Xhr method for posting store wall posts (xhr_addPinPost) and made it part of the REST API (POST on /api/stores/{id}/posts}. !1226 #719 @janopae
+- Removed mayLegacy function from session !1248 @alex.simm
 
 ## Dev/Test/CI stuff